requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cmake
cycler==0.10.0
dlib==19.7.0
Flask==1.1.2
Flask-Cors==3.0.10
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.4
numpy==1.19.5
opencv-python==4.5.1.48
Pillow==8.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.25.1
scipy==1.1.0
six==1.15.0
urllib3==1.26.3
Werkzeug==1.0.1
